# docs/api-specs/smart-inventory-openapi.yaml
openapi: 3.0.3
info:
  title: Smart Inventory Management API
  version: 1.0.0
  description: Multi-tenant SaaS API for restaurant inventory, purchasing, forecasting, waste, and profitability.
servers:
  - url: https://api.smart-inventory.example.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local Django API
security:
  - bearerAuth: []
...
paths:
...
  /forecasts/daily:
    get:
      tags: [Forecasts]
      summary: Get daily demand forecasts.
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
        - name: product_id
          in: query
          schema:
            type: string
            format: uuid
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Forecast list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Forecast'
  /recommendations/replenishment:
    get:
      tags: [Recommendations]
      summary: Get replenishment recommendations.
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
      responses:
        '200':
          description: Recommendations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReplenishmentRecommendation'
  /analytics/profitability:
    get:
      tags: [Analytics]
      summary: Get profitability analytics.
      parameters:
        - $ref: '#/components/parameters/RestaurantId'
        - name: from_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Profitability summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfitabilityReport'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RestaurantId:
      name: restaurant_id
      in: query
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Forecast:
      type: object
      properties:
        product_id:
          type: string
          format: uuid
        forecast_date:
          type: string
          format: date
        horizon_days:
          type: integer
        predicted_quantity:
          type: number
        confidence_low:
          type: number
        confidence_high:
          type: number
    ReplenishmentRecommendation:
      type: object
      properties:
        product_id:
          type: string
          format: uuid
        recommended_quantity:
          type: number
        supplier_id:
          type: string
          format: uuid
        reason:
          type: string
    ProfitabilityReport:
      type: object
      properties:
        restaurant_id:
          type: string
          format: uuid
        revenue:
          type: number
        food_cost:
          type: number
        waste_cost:
          type: number
        gross_margin:
          type: number
...